src/semiff/simulation/warp_env.py
# warp_env.py - Warp 场景搭建与参数辨识 (支持 3DGS 绑定)
import warp as wp
import numpy as np
import pickle
from pathlib import Path
from typing import Optional, Dict, List

try:
    from yourdfpy import URDF
except ImportError:
    print("Warning: yourdfpy not installed. Install with `pip install yourdfpy`")

from ..core.logger import get_logger

logger = get_logger(__name__)


class GaussianRobotRenderer:
    """
    基于 3DGS 绑定的机器人渲染器
    参考: real2sim-eval 的 Gaussian Binding 实现
    """

    def __init__(self, binding_data_path: str, urdf_path: str):
        """
        初始化渲染器

        Args:
            binding_data_path: 绑定数据文件路径 (step3_bind_robot_gs.py 的输出)
            urdf_path: URDF 文件路径
        """
        self.binding_data_path = Path(binding_data_path)
        self.urdf_path = urdf_path

        # 加载绑定数据
        self._load_binding_data()

        # 加载 URDF
        self.robot = URDF.load(urdf_path)

        # 缓存 link 变换
        self.link_transforms = {}

        logger.info(f"✅ Loaded Gaussian Robot Renderer")
        logger.info(f"   Gaussians: {self.num_gaussians}")
        logger.info(f"   Links: {len(self.link_names)}")

    def _load_binding_data(self):
        """加载绑定数据"""
        with open(self.binding_data_path, 'rb') as f:
            data = pickle.load(f)

        self.link_names = data['link_names']
        self.link_indices = data['link_indices']
        self.transform_matrix = data['transform_matrix']
        self.num_gaussians = data['num_gaussians']

        logger.info(f"Loaded binding data: {len(self.link_names)} links, {self.num_gaussians} Gaussians")

    def update_robot_pose(self, joint_angles: Dict[str, float]):
        """
        更新机器人姿态

        Args:
            joint_angles: 关节角度字典 {joint_name: angle}
        """
        self.robot.update_cfg(joint_angles)

        # 计算每个 Link 的变换矩阵
        self.link_transforms = {}
        for link_name in self.link_names:
            T_link = self.robot.get_transform(link_name)
            self.link_transforms[link_name] = T_link

    def get_gaussian_positions(self, canonical_positions: np.ndarray) -> np.ndarray:
        """
        根据当前机器人姿态计算高斯点位置

        Args:
            canonical_positions: 规范姿态下的高斯点位置 (来自 binding 数据)

        Returns:
            current_positions: 当前姿态下的高斯点位置
        """
        if not self.link_transforms:
            # 如果没有设置姿态，使用规范姿态
            return canonical_positions

        # 为每个高斯点计算当前位置
        current_positions = np.zeros_like(canonical_positions)

        for i, link_idx in enumerate(self.link_indices):
            if link_idx >= 0:
                # 该高斯点绑定到某个 Link
                link_name = self.link_names[link_idx]
                T_link = self.link_transforms[link_name]

                # 变换高斯点位置
                pos_homo = np.append(canonical_positions[i], 1.0)
                current_pos = T_link @ pos_homo
                current_positions[i] = current_pos[:3]
            else:
                # 未绑定的高斯点保持原位
                current_positions[i] = canonical_positions[i]

        return current_positions

    def render_frame(self, canonical_gaussians, camera_params):
        """
        渲染一帧 (这里是概念实现，实际需要集成 gsplat 或类似渲染器)

        Args:
            canonical_gaussians: 规范姿态下的高斯参数
            camera_params: 相机参数

        Returns:
            rendered_image: 渲染结果
        """
        # 获取当前高斯点位置
        current_positions = self.get_gaussian_positions(canonical_gaussians['positions'])

        # 更新高斯参数
        updated_gaussians = canonical_gaussians.copy()
        updated_gaussians['positions'] = current_positions

        # TODO: 调用实际的渲染器 (gsplat, etc.)
        # 这里只是概念代码
        rendered_image = self._conceptual_render(updated_gaussians, camera_params)

        return rendered_image

    def _conceptual_render(self, gaussians, camera_params):
        """概念渲染函数 (需要用实际的 3DGS 渲染器替换)"""
        # 这是一个占位符，实际实现需要:
        # 1. 集成 gsplat 或 gaussian-splatting 的渲染管道
        # 2. 设置相机参数
        # 3. 执行光栅化
        return np.zeros((480, 640, 3), dtype=np.uint8)


class WarpScene:
    """Warp 物理仿真场景"""

    def __init__(self, urdf_path: str, binding_data_path: Optional[str] = None):
        self.urdf_path = urdf_path
        self.binding_data_path = binding_data_path

        # 初始化 Warp
        wp.init()

        # 加载机器人
        self.robot = URDF.load(urdf_path)

        # 高斯渲染器 (如果提供绑定数据)
        self.gaussian_renderer = None
        if binding_data_path:
            self.gaussian_renderer = GaussianRobotRenderer(binding_data_path, urdf_path)

        # Warp 相关变量
        self.integrator = None
        self.renderer = None

    def setup_scene(self):
        """搭建物理场景"""
        # TODO: 实现完整的 Warp 场景搭建
        # 1. 创建刚体
        # 2. 设置关节约束
        # 3. 配置碰撞体
        # 4. 设置初始状态

        logger.info("Setting up Warp physics scene...")

        # 概念代码 - 需要根据实际需求实现
        # self.model = wp.Model()
        # self.state = self.model.state()

    def step_simulation(self, joint_targets: Optional[Dict[str, float]] = None):
        """执行一步仿真"""
        if joint_targets:
            # 设置关节目标
            pass

        # 执行物理步进
        # TODO: 实现 Warp 仿真步进

    def render_gaussians(self, canonical_gaussians, camera_params):
        """渲染高斯机器人"""
        if self.gaussian_renderer:
            return self.gaussian_renderer.render_frame(canonical_gaussians, camera_params)
        else:
            logger.warning("No Gaussian renderer available")
            return None


def setup_warp_scene(urdf_path, background_splat_path=None, binding_data_path=None):
    """
    搭建Warp物理场景 (增强版，支持 3DGS 绑定)

    Args:
        urdf_path: URDF文件路径
        background_splat_path: 背景Splat文件路径 (可选)
        binding_data_path: 高斯绑定数据路径 (可选)
    """
    scene = WarpScene(urdf_path, binding_data_path)
    scene.setup_scene()

    logger.info("✅ Warp scene setup completed")
    return scene


def run_simulation(scene: WarpScene, num_steps=1000, joint_trajectory=None):
    """
    运行物理仿真 (支持高斯渲染)

    Args:
        scene: WarpScene 实例
        num_steps: 仿真步数
        joint_trajectory: 关节轨迹 (可选)
    """
    logger.info(f"Running simulation for {num_steps} steps...")

    for step in range(num_steps):
        # 获取当前关节目标
        joint_targets = None
        if joint_trajectory and step < len(joint_trajectory):
            joint_targets = joint_trajectory[step]

        # 执行仿真步进
        scene.step_simulation(joint_targets)

        # 可以在这里添加渲染或数据记录
        if step % 100 == 0:
            logger.info(f"Step {step}/{num_steps}")

    logger.info("✅ Simulation completed")


def system_identification(sim_trajectory, real_trajectory):
    """
    系统辨识：优化物理参数以匹配真实轨迹

    Args:
        sim_trajectory: 仿真轨迹
        real_trajectory: 真实轨迹

    Returns:
        optimized_params: 优化后的物理参数
    """
    # TODO: 实现系统辨识算法
    # 可以使用梯度下降、进化算法等优化物理参数
    logger.info("Running system identification...")

    # 概念实现
    optimized_params = {
        'mass': 1.0,
        'friction': 0.5,
        'damping': 0.1
    }

    return optimized_params



